pyproject.toml
[tool.black]
line-length = 79

[tool.poetry]
name = "manage-chanell-bot"
version = "0.1.0"
description = ""
authors = ["Me"]
readme = "README.md"
packages = [{include = "manage_chanell_bot"}]

[tool.poetry.dependencies]
python = "^3.10"
aiogram = {version = "^3.0.0b1", allow-prereleases = true}
python-dotenv = "^0.21.0"
fluentogram = "^1.1.6"
sqlalchemy = "^2rc"
alembic = "^1.9.1"
asyncpg = "^0.27.0"
uvicorn = {extras = ["standart"], version = "^0.20.0"}
fastapi = "^0.89.1"

[tool.poetry.group.dev.dependencies]
flake8 = "^6.0.0"
isort = "^5.10.1"
pep8-naming = "^0.13.2"
black = "^22.10.0"
mypy = "^0.991"
flake8-bugbear = "^22.12.6"
flake8-builtins = "^2.0.1"
flake8-spellcheck = "^0.28.0"
flake8-comprehensions = "^3.10.1"
flake8-commas = "^2.1.0"
flake8-quotes = "^3.3.1"
pytest = "^7.2.0"
add-trailing-comma = "^2.4.0"
types-pytz = "^2022.7.0.0"
pytest-asyncio = "^0.20.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
